+++ src/pages/Home/Home.tsx
@@ -1,0 +1,15 @@
+import Carrosel from '../../components/Carrosel/Carrosel';
+import ProductList from '../../components/ProductList/ProductList';
+import PromoBanner from '../../components/PromoBanner/PromoBanner';
+
+function Home() {
+  return (
+    <>
+      <PromoBanner />
+      <Carrosel />
+      <ProductList />
+    </>
+  );
+}
+
+export default Home;
