+++ src/routes.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Header from './components/Header/Header';
 import Footer from './components/Footer/Footer';
-import Home from './pages/Home';
+import Home from './pages/Home/Home';
 import Sobre from './pages/Sobre/Sobre';
 import NotFound from './pages/NotFound/NotFound';
